Created AbstractDataAccess that DataAccess inherits from.  This lets us test
